fix: resolve compilation errors in PR #26
This commit fixes multiple compilation issues introduced in PR #26:

1. Type mismatch in gopclntab.go:163 - Cast uint32 to int64
2. Removed unused imports in tracee.go (helpers, utils packages)
3. Fixed function names in gopclntab_test.go to use correct Tracee prefix
4. Added //go:noinline directives to test functions to prevent optimization

All tests now pass successfully.

+++ pkg/trace/gopclntab.go
@@ -160,5 +160,5 @@ func (t *UserTracee) getSymbolOffset(symName string) (int64, error) {
 			Msg("failed to get function offset via helper")
 		return 0, err
 	}
-	return offset, nil
+	return int64(offset), nil
 }
